Move mockObjectComponent out of the two test files that use it, to prevent duplicate code
